Feat: 작성하기 후 돌아오는 오류 해결, 화살표 이동 기능 구현(숫자표시는 아직 에러)
