+++ src/Atoms/PageOffset.jsx
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+import { atom } from 'recoil';
+
+export const PageOffset = atom({
+  key: 'PageOffset',
+  default: 0,
+});
