+++ src/Pages/MainPage.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from 'react-router-dom';
 import instance from '../lib/Request';
 
 import ContentList from '../Components/MainPage/ContentList';
-import { ReactComponent as pageMovingBtn } from '../assests/pageMovingBtn.svg';
+import Pagination from '../Components/MainPage/Pagination';
 
 function MainPage() {
   const navigate = useNavigate();
@@ -12,7 +12,6 @@ function MainPage() {
     navigate('/writing');
   };
 
-  // 글 목록 불러오기
   const [sorting, setSorting] = useState('createAt');
   const onClickNew = () => {
     setSorting('createAt');
@@ -22,22 +21,21 @@ function MainPage() {
   };
 
   const [posts, setPosts] = useState([]);
-  const [totalPages, setTotalPages] = useState(0);
+  // const [maskingPosts, setMaskingPosts] = useState([]);
 
-  // 페이지네이션
-  const [currentPage, setCurrentPage] = useState(1);
-  const offset = (currentPage - 1) * 10;
-  const [currentPageArray, setCurrentPageArray] = useState([]);
-  const [totalArray, setTotalArray] = useState([]);
-
-  const sliceArray = (total, limit) => {
-    const totalPageArray = Array(total)
-      .fill()
-      .map((_, i) => i);
-    return Array(Math.ceil(total / limit))
-      .fill()
-      .map(() => totalPageArray.splice(0, limit));
-  };
+  // const maskingName = (name) => {
+  //   if (name.length > 2) {
+  //     const userName = name.split('');
+  //     userName.forEach((n, i) => {
+  //       if (i === 0) return;
+  //       userName[i] = '*';
+  //     });
+  //     const joinName = userName.join();
+  //     return joinName.replace(/,/g, '');
+  //   }
+  //   const pattern = /.$/;
+  //   return name.replace(pattern, '*');
+  // };
 
   const postAPI = async () => {
     try {
@@ -46,7 +44,6 @@ function MainPage() {
     } catch (err) {
       console.log('불러오기 실패!');
     }
-    setTotalPages(posts.length);
   };
 
   const sortAPI = async () => {
@@ -66,11 +63,8 @@ function MainPage() {
         console.log('불러오기 실패!');
       }
     }
-    setTotalPages(posts.length);
   };
 
-  // useEffect 무한렌더링 해결
-  // 최신순, 좋아요순 변경 시 렌더링
   useEffect(() => {
     postAPI();
   }, []);
@@ -78,22 +72,6 @@ function MainPage() {
   useEffect(() => {
     sortAPI();
   }, [sorting]);
-
-  useEffect(() => {
-    const slicedPageArray = sliceArray(totalPages, 5);
-
-    setTotalArray(slicedPageArray);
-    setCurrentPageArray(slicedPageArray[0]);
-
-    if (currentPage % 5 === 1) {
-      setCurrentPageArray(totalArray[Math.floor(currentPage / 5)]);
-    } else if (currentPage % 5 === 0) {
-      setCurrentPageArray(totalArray[Math.floor(currentPage / 5) - 1]);
-    }
-
-    console.log(currentPageArray);
-    console.log(totalArray);
-  }, [currentPage]);
 
   return (
     <MainWrap>
@@ -117,37 +95,10 @@ function MainPage() {
               <p>답변</p>
             </Header2ndContent>
           </ListHeader>
-          <ContentList posts={posts} totalPages={totalPages} offset={offset} />
+          <ContentList posts={posts} />
         </ListContainer>
         <WriteButton onClick={goWriting}>작성하기</WriteButton>
-        <ListPagesButton>
-          <MovingBtnWrap
-            onClick={() => setCurrentPage(currentPage - 1)}
-            disabled={currentPage === 1}
-          >
-            <MovingBtnStyle />
-          </MovingBtnWrap>
-          <ScrollWidth>
-            <PageBtnContainer>
-              {/* {currentPageArray.map((i) => (
-                <PageBtn
-                  key={i + 1}
-                  onClick={() => setCurrentPage(i + 1)}
-                  active={currentPage === i + 1}
-                >
-                  {i + 1}
-                </PageBtn>
-              ))} */}
-              <PageBtn />
-            </PageBtnContainer>
-          </ScrollWidth>
-          <MovingBtnWrap
-            onClick={() => setCurrentPage(currentPage + 1)}
-            disabled={currentPage === totalPages}
-          >
-            <MovingBtnStyle rightbtn="true" />
-          </MovingBtnWrap>
-        </ListPagesButton>
+        <Pagination posts={posts} />
       </MainContainer>
     </MainWrap>
   );
@@ -230,48 +181,6 @@ const Header2ndContent = styled.div`
     font-size: 25px;
     user-select: none;
   }
-`;
-
-const ListPagesButton = styled.div`
-  display: flex;
-  justify-content: space-between;
-  width: 350px;
-  height: 40px;
-  border: 1px solid #adadad;
-  border-radius: 5px;
-`;
-
-const MovingBtnWrap = styled.button`
-  border: none;
-  background-color: transparent;
-`;
-
-const MovingBtnStyle = styled(pageMovingBtn)`
-  transform: ${(props) => props.rightbtn && 'rotate(180deg)'};
-  cursor: pointer;
-`;
-
-const ScrollWidth = styled.div`
-  /* display: flex; */
-  /* justify-content: flex-start; */
-  /* width: 165px;
-  height: 40px; */
-`;
-
-const PageBtnContainer = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  user-select: none;
-`;
-
-const PageBtn = styled.div`
-  /* margin: 10px; */
-  margin: 5px 10px 5px 10px;
-  font-weight: ${(props) => (props.active ? '700' : '600')};
-  font-size: ${(props) => (props.active ? '28px' : '25px')};
-  color: ${(props) => (props.active ? '#0186D1' : 'none')};
-  cursor: pointer;
 `;
 
 const WriteButton = styled.button`
